Update cart display on adding products
The 'Add to Cart' functionality has been enhanced to immediately update the cart display. Each time a product is added to the cart, the incremented total count is rendered on the cart icon, improving user experience and immediate feedback. This is enabled by adding a 'cart-container' id to the cart div element.