Modify product replacement script in script.js
In this commit, we've modified an existing loop in script.js that changes certain product names. The modification enhances the previous replacements of any product title containing variations of "Кури" with "Индейка", now allowing for additional text to be retained during the replacement process. This offers more flexibility in altering displayed product names.